Handle .ekstazi folder missing after build

+++ ekstazi/src/main/java/com/pluralsight/ekstazi/EkstaziArtifactArchiver.java
@@ -21,27 +21,30 @@ public class EkstaziArtifactArchiver extends ArtifactArchiver {
     }
 
     public boolean perform(AbstractBuild build, Launcher launcher,
-            BuildListener listener) throws InterruptedException {
-        boolean result = false;
+            BuildListener listener) {
+        boolean result = true;
         try {
-        result = super.perform(build, launcher, listener);
-        FilePath workspacePath = build.getWorkspace();
-        FilePath ekstaziPath = workspacePath.child(".ekstazi");
-        ekstaziPath.deleteRecursive();
-        FilePath artifactsDir = new FilePath(build.getArtifactsDir());
-        FilePath buildDir = new FilePath(build.getProject().getBuildDir());
-        buildDir = buildDir.child("lastSuccessfulEkstaziBuild");
-        buildDir.symlinkTo(Integer.toString(build.number), listener);
-        listener.getLogger().println("Archiving Ekstazi results.");
+            //ttest here
+            FilePath workspacePath = build.getWorkspace();
+            FilePath ekstaziPath = workspacePath.child(".ekstazi");
+            if(ekstaziPath.exists()) {
+                result = super.perform(build, launcher, listener);
+                ekstaziPath.deleteRecursive();
+                FilePath buildDir = new FilePath(build.getProject().getBuildDir());
+                buildDir = buildDir.child("lastSuccessfulEkstaziBuild");
+                buildDir.symlinkTo(Integer.toString(build.number), listener);
+                listener.getLogger().println("Archiving Ekstazi results.");
+            }
         } catch (IOException | InterruptedException e) {
             listener.getLogger().println("Unable to archive old Ekstazi output.");
         }
         return result;
     }
+
     @Extension
     public static final class DescriptorImpl extends BuildStepDescriptor<Publisher> {
         public boolean isApplicable(Class<? extends AbstractProject> aClass) {
-                return true;
+            return true;
         }
 
         public String getDisplayName() {
